bug in tedt

+++ tests/test_io.py
@@ -250,8 +250,9 @@ def test_carriage_return_fasta(fasta_with_carriage_return_filename):
 
 # @pytest.mark.xfail
 def test_carriage_return_fai(fasta_with_carriage_return_filename):
-    # remove file
-    os.remove(fasta_with_carriage_return_filename + '.fai')
+    # remove file if it exists
+    if os.path.exists(fasta_with_carriage_return_filename + '.fai'):
+        os.remove(fasta_with_carriage_return_filename + '.fai')
     fai = bnp.open_indexed(fasta_with_carriage_return_filename)
     assert_encoded_array_equal(fai['test_sequence_id_here'].raw(), 'GACTG')
     assert_encoded_array_equal(fai['test_sequence_id_here2'].raw(), 'GACTCGAG')
